requirement.txt.txt
pandas==1.4.3

numpy==1.23.1

python-can==4.0.0

scikit-learn==1.1.2